Can start map from Quest now, rename HostLobbySetupPanel to LobbySetupPanel
