+++ MultiQuestensions/MultiQuestensions.vcxproj.filters
@@ -48,7 +48,7 @@
     <ClCompile Include="src\Beatmaps\PreviewBeatmapStub.cpp">
       <Filter>src\Beatmaps</Filter>
     </ClCompile>
-    <ClCompile Include="src\UI\HostLobbySetupPanel.cpp" />
+    <ClCompile Include="src\UI\LobbySetupPanel.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="include\main.hpp">
@@ -135,6 +135,6 @@
     <ClInclude Include="extern\songdownloader\shared\Types\BeatSaver\User.hpp">
       <Filter>extern</Filter>
     </ClInclude>
-    <ClInclude Include="include\UI\HostLobbySetupPanel.hpp" />
+    <ClInclude Include="include\UI\LobbySetupPanel.hpp" />
   </ItemGroup>
 </Project>
